Move insert frame from helpers to extends
